refactor: long으로 타입 변환

+++ src/main/java/step3/domain/PrizeCalculator.java
@@ -15,6 +15,6 @@ public class PrizeCalculator {
     }
 
     public long getResult() {
-        return winnerCount * winningType.getPrize();
+        return (long) winnerCount * winningType.getPrize();
     }
 }
